# docs/swagger.yaml
basePath: /api
definitions:
  rip_internal_app_ds.Route:
    properties:
      delay:
        type: integer
      description:
        type: string
      distance:
        type: integer
      imageURL:
        type: string
      routeID:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Route Information Platform API
  title: RIP API
  version: "1.0"
paths:
...
  /draft/addroute/{route_id}:
    post:
      description: Add route to draft. Create new darft if darft not existing. For
        authentificated users (client, moderator)
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add route to draft
      tags:
      - routes
  /routes:
    get:
      description: Get all routes list with optional filters
      parameters:
      - description: Minimum distance filter
        in: query
        name: min_distance
        type: integer
      - description: Maximum distance filter
        in: query
        name: max_distance
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: routes
          schema:
            items:
              $ref: '#/definitions/rip_internal_app_ds.Route'
            type: array
      summary: Get all routes
      tags:
      - routes
    post:
      consumes:
      - application/json
      description: Create new route. Only for moderator
      produces:
      - application/json
      responses:
        "200":
          description: route
          schema:
            $ref: '#/definitions/rip_internal_app_ds.Route'
      security:
      - BearerAuth: []
      summary: Create new route
      tags:
      - routes
  /routes/{route_id}:
    delete:
      description: Delete existing route. Only for moderator
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Маршрут успешно удален
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete existing route
      tags:
      - routes
    get:
      description: Get route information by ID
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: route
          schema:
            $ref: '#/definitions/rip_internal_app_ds.Route'
      summary: Get route by ID
      tags:
      - routes
    put:
      consumes:
      - application/json
      description: Update existing route. Only for moderator
      produces:
      - application/json
      responses:
        "200":
          description: route
          schema:
            $ref: '#/definitions/rip_internal_app_ds.Route'
      security:
      - BearerAuth: []
      summary: Update existing route
      tags:
      - routes
  /routes/{route_id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload route image. Add new image if route doesn`t have one or
        change image. Only for moderator
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Изображение успешно загружено и обновлено
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upload route image
      tags:
      - routes
  /routespeedrequests:
    delete:
      consumes:
      - application/json
      description: Remove route from speed request. For authentificated users only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove route speed request
      tags:
      - routespeedrequests
    put:
      consumes:
      - application/json
      description: Upadet field arrival date in route speed request. For authentificated
        users only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update route speed request
      tags:
      - routespeedrequests
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"